+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.16.0] - 03-Apr-2021
+
+### Added
+
+- The `Text` widget has a new option `MaxTextCells` which can be used to limit
+  the maximum number of cells the widget keeps in memory.
+
+### Changed
+
+- Bump github.com/mattn/go-runewidth from 0.0.10 to 0.0.12.
+
 ## [0.15.0] - 06-Mar-2021
 
 ### Changed
@@ -453,7 +464,8 @@ identifiers shouldn't be used externally.
 - The Gauge widget.
 - The Text widget.
 
-[unreleased]: https://github.com/mum4k/termdash/compare/v0.15.0...devel
+[unreleased]: https://github.com/mum4k/termdash/compare/v0.16.0...devel
+[0.16.0]: https://github.com/mum4k/termdash/compare/v0.15.0...v0.16.0
 [0.15.0]: https://github.com/mum4k/termdash/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/mum4k/termdash/compare/v0.13.0...v0.14.0
 [0.13.0]: https://github.com/mum4k/termdash/compare/v0.12.2...v0.13.0